torchseg/requirements.txt
numpy==1.19.2
packaging==20.4
torch==1.7.0
pretrainedmodels==0.7.4
tabulate==0.8.7
segmentation_models_pytorch==0.1.2
tqdm==4.50.2
netpbmfile==2020.10.18
matplotlib==3.3.3
pandas==1.1.3
tensorflow==2.2.0
torchvision==0.8.1
easydict==1.9
imgaug==0.4.0
fire==0.3.1
Pillow==8.0.1
PyYAML==5.3.1
tensorboardX==2.1
